moved entity/dto converting functionality to specified classes
